Tests for old world behaviour pre ref-counting
The default CF is OK in the old world because it just gets disowned on creation. Of course we will crap out when using it with a closed DB, which if we structure the new world properly will be prevented.

Accessing a closed DB (by copying it out of the AutoCloseable scope, and using it after) causes a crash in the old world, as we expect. We expect the new world to hand us a nice Java exception.