Add benchmarks to CircleCI runner
Build a java release on the configured CircleCI runner, and then run a couple of tests with that and capture the output.

Main complexity is ensuring that the runner machine (Amazon Liux 2) has the right software and libs; the setup is documented elsewhere.

Some of the software is installed through local builds, so we need /usr/local/lib in the include paths.

We are unable to upload artifacts from the runner due to permission issues. We’ve tried a lot of ways to fix this. Workaround at present is to output logs as MIME.