Rethink. patch original RocksDB as RocksNative
Add some C++ wrapper to turn the C++ handle into a C++ RCA (Reference Counted API) object.
Pull on the cascading changes necessary.
Nearly in a position to run RefCountTest in the new world, at least for opening up a DB.
And remove all the shadow (org.rocksdb.api package) code as that path isn't the one we need to take.